backend/requirements.txt
fastapi
uvicorn
openai
langchain
langchain-openai
faiss-cpu
python-docx
PyPDF2
python-dotenv